Upgraded to LevelDB 1.16
 * fixed autocompact test ratio

+++ util/cache_test.cc
@@ -11,9 +11,9 @@
 namespace leveldb {
 
 // Conversions between numeric keys/values and the types expected by Cache.
-static std::string EncodeKey(int k) {
+static std::string EncodeKey(intptr_t k) {
   std::string result;
-  PutFixed32(&result, k);
+  PutFixed32(&result, static_cast<uint32_t>(k));
   return result;
 }
 static int DecodeKey(const Slice& k) {
@@ -45,7 +45,7 @@ class CacheTest {
     delete cache_;
   }
 
-  int Lookup(int key) {
+  intptr_t Lookup(int key) {
     Cache::Handle* handle = cache_->Lookup(EncodeKey(key));
     const intptr_t r = (handle == NULL) ? -1 : DecodeValue(cache_->Value(handle));
     if (handle != NULL) {
@@ -161,10 +161,10 @@ TEST(CacheTest, HeavyEntries) {
     index++;
   }
 
-  int cached_weight = 0;
+  intptr_t cached_weight = 0;
   for (int i = 0; i < index; i++) {
-    const int weight = (i & 1 ? kLight : kHeavy);
-    int r = Lookup(i);
+    const intptr_t weight = (i & 1 ? kLight : kHeavy);
+    intptr_t r = Lookup(i);
     if (r >= 0) {
       cached_weight += weight;
       ASSERT_EQ(1000+i, r);
